move time handlers into new files, make sure handler map doesn't cause crash, fix round transition logic
