+++ packages/server/src/index.ts
@@ -18,7 +18,7 @@ import { Strategy as LocalStrategy } from "passport-local";
 import { UserResponse } from "@ogfcommunity/variants-shared";
 import { router as apiRouter } from "./api";
 import * as socket_io from "./socket_io";
-import { TimeoutService } from "./timeout";
+import { TimeoutService } from "./time-control/timeout";
 
 const LOCAL_ORIGIN = "http://localhost:5173";
